ecommerce_corporate_account: bug fixing
git-svn-id: https://silverstripe-ecommerce.googlecode.com/svn/modules/ecommerce_corporate_account/trunk@1551 950b781c-c544-efab-2f99-90083ee69683

+++ javascript/AddUpProductsToOrderPage.js
@@ -75,7 +75,7 @@ var AddUpProductsToOrderPage = {
 				jQuery(".qty input").live(
 					"change",
 					function(){
-						var val = parseInt(jQuery(this).val());
+						var val = parseFloat(jQuery(this).val());
 						if(!val) {
 							jQuery(this).val(0);
 							jQuery(this).addClass("toBeCompleted");
@@ -144,7 +144,7 @@ var AddUpProductsToOrderPage = {
 	updateRows: function ()  {
 		for(var i = 0; i < this.rowNumbers; i++) {
 			var price = parseFloat(jQuery("#Buyable_"+i+" select option:selected").attr("rel"));
-			var qty = parseInt(jQuery("#Qty_"+i+" input").val());
+			var qty = parseFloat(jQuery("#Qty_"+i+" input").val());
 			var total = Math.round((qty * price * 100))/100;
 			if(total && total != NaN && total > 0) {
 				jQuery("#Total_"+i).text("$" +total)
